fix: break internal-metadata <-> migration circular dep + clarify env:check description
1. Extracted all migration error classes (MigrationError,
   IrreversibleMigration, DuplicateMigrationVersionError,
   DuplicateMigrationNameError, UnknownMigrationVersionError,
   IllegalMigrationNameError, InvalidMigrationTimestampError,
   PendingMigrationError, ConcurrentMigrationError,
   NoEnvironmentInSchemaError, ProtectedEnvironmentError,
   EnvironmentMismatchError, EnvironmentStorageError) into a new
   standalone module: packages/activerecord/src/migration-errors.ts.
   migration.ts now re-exports them for existing consumers, so the
   public API is unchanged.

   InternalMetadata can now statically import EnvironmentStorageError
   from migration-errors.ts instead of dynamic-importing it through
   migration.ts. The previous `await import('./migration.js')` in
   InternalMetadata.set() was a lazy workaround for the circular
   dependency — migration.ts imports InternalMetadata, and if a
   disabled InternalMetadata.set() fired during module initialization
   it could see partially-initialized exports.

2. trails db environment:check description now calls out both failure
   modes — protected-env stamp AND EnvironmentMismatchError when
   stored != current env. The previous description only mentioned the
   protected case.
